moved and renamed author information filter
